REST-test task completed
