+++ HW_Code/HW-REST/Helpers/counter.js
@@ -1,0 +1,10 @@
+let counter = function(array) {
+  let n = 0;
+  for (let i = 0; i < array.length; i++) {
+    n += 1;
+    return n;
+  }
+  return n;
+};
+
+export default counter();
