+++ package.json
@@ -9,7 +9,7 @@
     "test": "protractor conf.js",
     "testcuc": "protractor cucumber-conf.js",
     "rest-test-cuc": "cucumber-rest-conf.js",
-    "rest-mocha": "mocha"
+    "test-mocha": "mocha './HW_Code/HW-REST/test/test.js'"
   },
   "author": "",
   "license": "ISC",
@@ -24,6 +24,7 @@
     "eslint-config-prettier": "^6.7.0",
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-protractor": "^2.1.1",
+    "expect.js": "^0.3.1",
     "npm": "^6.13.4",
     "prettier": "^1.19.1",
     "protractor": "^5.4.2",
